requirements.txt
blinker==1.9.0
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mqtt==0.0.1
paho-mqtt==2.1.0
PyMySQL==1.1.1
Werkzeug==3.1.3
